requirements.txt
Flask==3.0.3
supabase==2.4.5
python-dotenv==1.0.1
gunicorn==22.0.0
